# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/orders": {
      "post": {
        "operationId": "OrderController_open",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ordem criada com sucesso"
          }
        },
        "summary": "Abre uma nova ordem de servico",
        "tags": [
          "orders"
        ]
      },
      "get": {
        "operationId": "OrderController_list",
        "parameters": [
          {
            "name": "includeClosed",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista retornada com sucesso"
          }
        },
        "summary": "Lista ordens com filtros opcionais",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/queue": {
      "get": {
        "operationId": "ExecutionController_getQueue",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lista fila de execucao das ordens",
        "tags": [
          "execution"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrderController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consulta detalhes de uma ordem",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/history": {
      "get": {
        "operationId": "OrderController_getHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consulta historico de status da ordem",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "operationId": "OrderController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Atualiza status da ordem",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/items": {
      "post": {
        "operationId": "OrderController_addItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Adiciona item em uma ordem",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/items/{itemId}": {
      "delete": {
        "operationId": "OrderController_removeItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Remove item da ordem",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/cancel": {
      "patch": {
        "operationId": "OrderController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Cancela uma ordem",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/budget": {
      "post": {
        "operationId": "BudgetController_generate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateBudgetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Budget gerado com sucesso"
          }
        },
        "summary": "Gera ou atualiza budget da ordem",
        "tags": [
          "budget"
        ]
      },
      "get": {
        "operationId": "BudgetController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consulta budget da ordem",
        "tags": [
          "budget"
        ]
      }
    },
    "/orders/{id}/budget/approve": {
      "post": {
        "operationId": "BudgetController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Aprova budget da ordem",
        "tags": [
          "budget"
        ]
      }
    },
    "/orders/{id}/budget/reject": {
      "post": {
        "operationId": "BudgetController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectBudgetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Rejeita budget da ordem",
        "tags": [
          "budget"
        ]
      }
    },
    "/catalog/services": {
      "get": {
        "operationId": "CatalogController_listServices",
        "parameters": [
          {
            "name": "includeInactive",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lista servicos do catalogo",
        "tags": [
          "catalog"
        ]
      },
      "post": {
        "operationId": "CatalogController_createService",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Cria servico no catalogo",
        "tags": [
          "catalog"
        ]
      }
    },
    "/catalog/services/{id}": {
      "put": {
        "operationId": "CatalogController_updateService",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Atualiza servico do catalogo",
        "tags": [
          "catalog"
        ]
      },
      "delete": {
        "operationId": "CatalogController_deactivateService",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Desativa servico do catalogo",
        "tags": [
          "catalog"
        ]
      }
    },
    "/catalog/parts": {
      "get": {
        "operationId": "CatalogController_listParts",
        "parameters": [
          {
            "name": "includeInactive",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lowStockOnly",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lista pecas do catalogo",
        "tags": [
          "catalog"
        ]
      },
      "post": {
        "operationId": "CatalogController_createPart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Cria peca no catalogo",
        "tags": [
          "catalog"
        ]
      }
    },
    "/catalog/parts/{id}": {
      "put": {
        "operationId": "CatalogController_updatePart",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Atualiza peca do catalogo",
        "tags": [
          "catalog"
        ]
      },
      "delete": {
        "operationId": "CatalogController_deactivatePart",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Desativa peca do catalogo",
        "tags": [
          "catalog"
        ]
      }
    },
    "/catalog/parts/{id}/stock": {
      "patch": {
        "operationId": "CatalogController_adjustStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustStockDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Ajusta estoque de peca",
        "tags": [
          "catalog"
        ]
      }
    },
    "/orders/{id}/execution": {
      "patch": {
        "operationId": "ExecutionController_updateStage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExecutionStageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Atualiza etapa de execucao da ordem",
        "tags": [
          "execution"
        ]
      }
    },
    "/orders/{id}/execution/complete": {
      "post": {
        "operationId": "ExecutionController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Conclui execucao da ordem e move para pagamento",
        "tags": [
          "execution"
        ]
      }
    }
  },
  "info": {
    "title": "Order Service API",
    "description": "API de ciclo de vida de ordem de servico, catalogo, budget e execucao",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "OpenOrderDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {}
      },
      "AddItemDto": {
        "type": "object",
        "properties": {}
      },
      "GenerateBudgetDto": {
        "type": "object",
        "properties": {}
      },
      "RejectBudgetDto": {
        "type": "object",
        "properties": {}
      },
      "CreateServiceDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePartDto": {
        "type": "object",
        "properties": {}
      },
      "AdjustStockDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateExecutionStageDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}